Time: 100 ms (64.00%), Space: 9.4 MB (54.77%) - LeetHub

+++ 823-binary-trees-with-factors/823-binary-trees-with-factors.cpp
@@ -32,9 +32,9 @@ public:
             mp[arr[i]] += kount;
         }
 
-        long long int ans = 0;
+        int ans = 0;
         for (int num : arr) ans = add(ans, mp[num]);
 
-        return (int) ans;
+        return ans;
     }
 };
